dynamic_admin_site/requirements.txt
django-solo==2.2.0
